Add `link_to` filter for easy linking
The `link_to` filter links to a page in the current language; if that isn't
available it falls back to linking to the page in the site's default language.

Example:
```
{{ 'About' | link_to('/about/') }}
```

Gives this when linking from an `en-gb` page:
```
<a href="/en-gb/about/">About</a>
```

Can be used with the `i81n` filter:
```
{{ 'navigation.about' | i18n | link_to('/about/') }}
```

Respects the i18n filter's use of the `lang` attribute around fallbacks - moves
the `lang` attribute to the link element if needed.