🗺 Update translation helper to be less hacky
Instead of checking the key against the default locale, which felt a bit hacky
and meant the use of an `if` statement, this now uses a test instance of i18next
that turns off the fallback language. Asking for the translation with fallbacks
turned off returns the key when there isn't a translation - so there can be a
check to see if the translation matches the key.