🚛 Add get helper for querying pages
This will query a single page and return the parameter for the language of the
page that the filter is used on - if used on a Spanish page, it'll check for the
Spanish version of the page being queried. If that isn't present, this will
fall back to the site's default language.