🇺🇳 Add markdown plugin for l10n links
This plugin lets you point to the a page without worrying about whether or not
that page has a translation available.

For example, on a Spanish page using the markdown `[homepage](/)` will check if
the root page is available in Spanish. If it is available then that page will
be linked to; if not, then the default language page will be linked to and a
`hreflang` attribute will be added to the link.