👩‍🔬 Update placeholder localisation tests
This needs a slightly different way of running locally, since the local `site`
folder could have the site from the previous build of the site - so it needs to
be deleted between runs.

`npm t` removes the site folder, then runs builds the site using the placeholder
pages, then runs the tests.

`npm run test` runs the tests assuming no previous build - so suitable for
running in a GitHub Action.

The workflows needed updating to add the `ELEVENTY_PLACEHOLDER` flag so that the
build uses the `_pages-placeholder` directory.