🎨 Make Aussie Flat UI colours available as custom properties
The idea being to keep the design's colour scheme monochrome with one key
colour - which makes it easier for a visitor to set a colour as a preference
whilst maintaining accessible colour contrast.

Adding this colour scheme means I can quickly change the `--key-colour` property
to a nice colour to check that what works in one colour works in another colour.